[ADD] refactoring
