+++ Cargo.toml
@@ -4,11 +4,19 @@ version = "0.1.0"
 edition = "2021"
 
 
+[lib]
+path = "src/lib.rs"
+
+[[bin]]
+path = "src/main.rs"
+name = "blog_backend"
+
 [dependencies]
 axum = { version = "0.6.1" }
 chrono = { version = "0.4.23" }
 dotenv = { version = "0.15.0" }
 jsonwebtoken = { version = "8.2.0" }
+reqwest = "0.11"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = { version = "1.0.91" }
 sqlx = { version = "0.6", features = [ "runtime-tokio-rustls", "migrate", "postgres" ] }
